µC-only: wire most of cartridge to µC
This wires pretty much everything but the DMA lines to the
microcontroller.  That's still a bit too far for me :)  Turns out I'm
not going to need combinatoric logic because the PLA already
decodes #IOx and #ROMx only when the CPU's AEC line is enable -- i.e.
when it's the CPU's turn for the bus.

I mostly grouped the outputs-toward-the-Commodore in the low bits of
port B, and the inputs-from-the-Commodore in the high bits.  Since this
allows the microcontroller to control #GAME and #EXROM dynamically, as
well as send interrupts and NMIs, I will hopefully be able to emulate as
many existing peripherals as possible.

The resistors on the data bus are currently a bit of CYA, based on the
EasyFlash design.  I'm going to leave it in for now, and I intend to use
them while breadboard-testing, but none of the other cartridges seem to
put them in there.