firmware: return data from a Frogger cartridge
I was using the data from `Frogger_(Parker_Brothers).crt`, but that will
need to be obtained separately.

It seems that doing memory lookups in the critical section is just a bit
too slow to reliably get data to the Commodore -- I observe at least one
bit error about 10% of the time.

Initially, I was just indexing into the FROGGER object.  Thinking that
the flash wait-states were causing my issues, I added the cache on the
stack.  As far as I can tell, rustc actually optimized that out, which
is how the read_volatile() came to be.

Unfortunately, none of those optimizations were enough to reliably read
the CBM80 header.