firmware: shave off one branch from the inner loop
This is actually enough to reliably read bytes from the Frogger image!

It takes long enough at boot that it doesn't immediately trip the
Commodore's cartridge boot mechanism, but this is enough to launch into
a green screen upon a SYS64738 command.  It still does not seem to
actually launch the game, though.